+++ vtr-ui/src/app/components/ui/ui-test-result-icon/ui-test-result-icon.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { UiTestResultIconComponent } from './ui-test-result-icon.component';
-import 'jasmine';
+// import 'jasmine';
 
 describe('UiTestResultIconComponent', () => {
 	let component: UiTestResultIconComponent;